Fix for https://jira.spectralogic.com/browse/BPBROWSER-63
If a file has no dacl entries, as may happen on a network-mounted file system, there wont be a principal entry.  A principal is a combination of
security provider, like NT AUTHORITY, and user name, e.g. NT AUTHORITY\Gracie.  This code is looking for the user name -- the second half of the
principal.  With no principal, there is no second half of the principal.